Offline reading and more labels
Using tshock's static db instance (I don't think that should be an issue but not 100% sure)

Endpoint for ssc server. Set in index.php to check for url options, and if player is offline, attempt to get the inventory as a rest response. Could definitely be better but I hardly know C# or programming that well, just editing and matching syntax & pattern from other plugins.